TODO 추가 기능 리액트로 리팩토링
