+++ docs/src/contexts/date_context.js
@@ -2,4 +2,4 @@
 import { createContext } from "react";
 
 export const MonthContext = createContext();
-export const SelectedDayContext = createContext();
+export const SelectedDayContext = createContext(new Date());
