+++ docs/src/contexts/todoList_context.js
@@ -1,0 +1,4 @@
+
+import { createContext } from "react";
+
+export const todoArrayConext = createContext();
