+++ package.json
@@ -75,6 +75,7 @@
     "serialize-javascript": "^6.0.2",
     "source-map": "^0.6.1",
     "source-map-support": "^0.5.21",
+    "styled-components": "^6.1.8",
     "supports-color": "^8.1.1",
     "tapable": "^2.2.1",
     "terser": "^5.27.0",
